refactor(ai/review): split monolithic chunker workflow module
Decompose the oversized `chunker/workflow_scripts.py` (1016 LOC, previously
module-size baselined) into cohesive leaf modules with zero behaviour change:

- `shell_run_parse.py` (694 LOC): shell tokenizing/stripping — interpreter,
  `uv run`, `env`/`VAR=`, `sudo`/`timeout`, `exec`/`command` wrappers, `-c`/`-e`
  inline-code payloads, and `cd` cwd tracking.
- `github_action_paths.py` (92 LOC): local `.github/actions/` root resolution
  for `uses:` matching.
- `workflow_scripts.py` (280 LOC): now only the YAML workflow matching layer
  (`run:`/`uses:` steps, multi-line/folded blocks), importing the two leaves.

The public API (`chunk_review_context`) and the `chunker.grouping` import seam
(`_is_workflow_linked_script`, `_script_referenced_in_workflow`,
`_workflow_text_for_matching`) are unchanged. All three new modules are under the
800-line gate, so the baseline entry is removed from `pyproject.toml` and
`lintro/utils/module_size.py` (no new baselines added).

Adds `test_chunker_import_compat.py` pinning the public and internal import
paths. Existing suite passes unchanged.

Closes #1024
